# src/main/resources/static/openapi.yml
openapi: 3.0.0
info:
  title: Gym CRM API
  version: 1.0.0
  description: API for managing gym trainees, trainers, training sessions, and user authentication.
...
paths:
...
  /api/v1/trainee/{username}/trainers:
    put:
      summary: Update trainee's trainers list
      description: Update the list of trainers assigned to a trainee
      tags:
        - Trainee Controller
      parameters:
        - name: Username
          in: header
          description: Trainee's username for authentication
          required: true
          schema:
            type: string
          example: yevhenii.kashtalian
        - name: Password
          in: header
          description: Trainee's password for authentication
          required: true
          schema:
            type: string
          example: b9T(ac^ydY
        - name: username
          in: path
          description: Username of the trainee to update trainers list
          required: true
          schema:
            type: string
          example: john.doe
      requestBody:
        description: List of trainers to assign to the trainee
        required: true
        content:
          application/json:
            example:
              trainers:
                - john.doe
      responses:
        '200':
          description: Successfully updated trainers list
          content:
            application/json:
              example:
                - username: john.doe
                  firstName: John
                  lastName: Doe
                  specializationId: 1
        '400':
          description: Invalid request body
          content:
            application/json:
              example:
                localDateTime: "2024-08-05T16:16:53.8490207"
                errorMessage: Failed to update trainee's trainers list, the request contains invalid fields
                errors:
                  - field: trainers
                    message: Trainers list cannot be empty
                    rejectedValue: [ ]
        '401':
          description: Missing authentication headers response
          content:
            application/json:
              examples:
                Missing headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Missing authentication headers
                Invalid headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Invalid credentials
        '404':
          description: Trainee not found
          content: { }
...
  /api/v1/trainer/{username}:
    get:
      summary: Get trainer by username
      description: Fetch trainer details by username
      tags:
        - Trainer Controller
      parameters:
        - name: Username
          in: header
          description: Trainer's username for authentication
          required: true
          schema:
            type: string
          example: john.doe
        - name: Password
          in: header
          description: Trainer's password for authentication
          required: true
          schema:
            type: string
          example: password123
        - name: username
          in: path
          description: Username of the trainer to fetch
          required: true
          schema:
            type: string
          example: john.doe
      responses:
        '200':
          description: Successfully retrieved trainer profile
          content:
            application/json:
              example:
                firstName: John
                lastName: Doe
                specializationId: "3"
                isActive: true
                trainees: [ ]
        '401':
          description: Missing authentication headers response
          content:
            application/json:
              examples:
                Missing headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Missing authentication headers
                Invalid headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Invalid credentials
        '404':
          description: Trainer not found
          content: { }

    put:
      summary: Update trainer details
      description: Update trainer information
      tags:
        - Trainer Controller
      parameters:
        - name: Username
          in: header
          description: Trainer's username for authentication
          required: true
          schema:
            type: string
          example: john.doe
        - name: Password
          in: header
          description: Trainer's password for authentication
          required: true
          schema:
            type: string
          example: password123
        - name: username
          in: path
          description: Username of the trainer to update
          required: true
          schema:
            type: string
          example: john.doe
      requestBody:
        description: Trainer details to be updated
        required: true
        content:
          application/json:
            example:
              firstName: John
              lastName: Doe
              specialization: "3"
              active: false
      responses:
        '200':
          description: Successfully updated trainer profile
          content:
            application/json:
              example:
                username: john.doe
                firstName: John
                lastName: Doe
                specialization: "3"
                isActive: true
        '400':
          description: Invalid request body
          content:
            application/json:
              example:
                localDateTime: "2024-08-05T16:16:53.8490207"
                errorMessage: Failed to update trainer, the request contains invalid fields
                errors:
                  - field: firstName
                    message: First name cannot be empty
                    rejectedValue: ""
        '401':
          description: Missing authentication headers response
          content:
            application/json:
              examples:
                Missing headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Missing authentication headers
                Invalid headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Invalid credentials
        '404':
          description: Trainer not found
          content: { }
  /api/v1/trainer:
    post:
      summary: Register a new trainer
      description: Adds a new trainer to the system
      tags:
        - Trainer Controller
      requestBody:
        description: Trainer details to be registered
        required: true
        content:
          application/json:
            example:
              firstName: John
              lastName: Doe
              specializationId: "2"
      responses:
        '200':
          description: Successfully registered trainer
          content:
            application/json:
              example:
                username: john.doe
                password: password123
        '400':
          description: Invalid request body
          content:
            application/json:
              example:
                localDateTime: "2024-08-05T16:16:53.8490207"
                errorMessage: Failed to register trainer, the request contains invalid fields
                errors:
                  - field: firstName
                    message: First name cannot be empty
                    rejectedValue: ""
  /api/v1/trainers/unassigned:
    get:
      summary: Get unassigned trainers
      description: Fetch a list of trainers who are not assigned to any trainee
      tags:
        - Trainer Controller
      parameters:
        - name: Username
          in: header
          description: Trainer's username for authentication
          required: true
          schema:
            type: string
          example: john.doe
        - name: Password
          in: header
          description: Trainer's password for authentication
          required: true
          schema:
            type: string
          example: password123
        - name: username
          in: query
          description: Trainee's username to fetch unassigned trainers
          required: true
          schema:
            type: string
          example: john.doe
      responses:
        '200':
          description: Successfully retrieved unassigned trainers
          content:
            application/json:
              example:
                - username: jane.doe
                  firstName: Jane
                  lastName: Doe
                  specializationId: "3"
        '401':
          description: Missing authentication headers response
          content:
            application/json:
              examples:
                Missing headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Missing authentication headers
                Invalid headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Invalid credentials
  /api/v1/trainer/{username}/trainings:
    get:
      summary: Get trainer's trainings
      description: Fetch all training sessions of a trainer
      tags:
        - Trainer Controller
      parameters:
        - name: Username
          in: header
          description: Trainer's username for authentication
          required: true
          schema:
            type: string
          example: john.doe
        - name: Password
          in: header
          description: Trainer's password for authentication
          required: true
          schema:
            type: string
          example: password123
        - name: username
          in: path
          description: Username of the trainer to fetch trainings
          required: true
          schema:
            type: string
          example: john.doe
        - name: from
          in: query
          description: Start date for filtering trainings (yyyy-MM-dd)
          required: false
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: to
          in: query
          description: End date for filtering trainings (yyyy-MM-dd)
          required: false
          schema:
            type: string
            format: date
          example: "2024-12-31"
        - name: traineeName
          in: query
          description: Trainee's name for filtering trainings
          required: false
          schema:
            type: string
          example: Jane Doe
      responses:
        '200':
          description: Successfully retrieved trainer's trainings
          content:
            application/json:
              example:
                - trainingName: Morning Cardio
                  trainingDate: "2024-12-01"
                  trainingType:
                    - name: CARDIO
                      id: "3"
                  trainingDuration: 60
                  traineeName: Jane Doe
        '401':
          description: Missing authentication headers response
          content:
            application/json:
              examples:
                Missing headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Missing authentication headers
                Invalid headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Invalid credentials
        '404':
          description: Trainer not found
          content: { }
  /api/v1/trainer/{username}/status:
    patch:
      summary: Change trainer's active status
      description: Toggle the active status of a trainer
      tags:
        - Trainer Controller
      parameters:
        - name: Username
          in: header
          description: Trainer's username for authentication
          required: true
          schema:
            type: string
          example: john.doe
        - name: Password
          in: header
          description: Trainer's password for authentication
          required: true
          schema:
            type: string
          example: password123
        - name: username
          in: path
          description: Username of the trainer to change status
          required: true
          schema:
            type: string
          example: john.doe
      responses:
        '200':
          description: Successfully changed trainer's status
        '401':
          description: Missing authentication headers response
          content:
            application/json:
              examples:
                Missing headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Missing authentication headers
                Invalid headers:
                  value:
                    localDateTime: "2024-08-05T16:16:53.8490207"
                    errorMessage: Invalid credentials
        '404':
          description: Trainer not found
          content: { }
...